+++ javascript/linked-list/__tests__/linked-list.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Require our linked list implementation
-const LinkedLilst = require('../index');
+const LinkedList = require('../index');
 
 describe('Linked List', () => {
   it('works', () => {
